[CREATE] OSLog
