+++ Platform/Sources/BamBooSetting/Color/BamBooColor.swift
@@ -2,5 +2,12 @@ import UIKit
 import UIUtil
 
 public extension UIColor{
+    static func rgb(red: CGFloat ,green: CGFloat,blue:CGFloat) -> UIColor{
+        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: 1)
+    }
+    static func rgba(red: CGFloat ,green: CGFloat,blue:CGFloat, alpha:CGFloat) -> UIColor{
+        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: alpha)
+    }
+    
     static let bambooGreen = rgb(red: 87, green: 204, blue: 77)
 }
