+++ Platform/Sources/UIUtil/CALayer/CALayerExt.swift
@@ -1,7 +1,7 @@
 import UIKit
 
 //MARK: - Shadow extension
-extension CALayer {
+public extension CALayer {
 func applySketchShadow(color: UIColor = .black,alpha: Float = 0.5,x:CGFloat,y:CGFloat,blur: CGFloat,spread: CGFloat = 0){
         shadowColor = color.cgColor
         shadowOpacity = alpha
