--- Platform/Sources/UIUtil/Color/ColorExt.swift
@@ -1,9 +1,0 @@
-import UIKit
-public extension UIColor{
-    static func rgb(red: CGFloat ,green: CGFloat,blue:CGFloat) -> UIColor{
-        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: 1)
-    }
-    static func rgba(red: CGFloat ,green: CGFloat,blue:CGFloat, alpha:CGFloat) -> UIColor{
-        return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: alpha)
-    }
-}
